Add read_file module for reading file contents
This commit adds the read_file.py module, which contains the read_file function for reading the contents of a file. The function takes a filename as input and returns the contents of the file as a string. This module will be used for reading log files in the log analysis project.